Implement timeoffset property in RPC GetInfo
Add GetMedianTimeOffset extension for Nodes collection
RPC GetInfo uses ConnectionManager.NodesCollection.GetMedianTimeOffset()
Fix some style violations in ConnectionManager

+++ Stratis.Bitcoin/RPC/Models/GetInfoModel.cs
@@ -24,7 +24,7 @@ namespace Stratis.Bitcoin.RPC.Models
         public int blocks { get; set; }
 
         [JsonProperty(Order = 5)]
-        public double timeoffset { get; set; }
+        public long timeoffset { get; set; }
 
         [JsonProperty(Order = 6, DefaultValueHandling = DefaultValueHandling.Ignore)]
         public int? connections { get; set; }
